feat(comms):[CDM-44] add backend push notification logic

+++ services/communication-service/src/main/java/ie/tcd/cdm/communication_service/dto/CreateNotificationDTO.java
@@ -10,9 +10,9 @@ public record CreateNotificationDTO(
         @NotNull
         String messageHeader,
         @NotNull
-        int messageSender,
+        long messageSender,
         @NotNull
-        int receiver,
+        long receiver,
         @NotNull
         String receiverGroup,
         @NotNull
