+++ services/communication-service/src/main/java/ie/tcd/cdm/communication_service/dto/UpdateNotificationDTO.java
@@ -9,9 +9,9 @@ public record UpdateNotificationDTO(
 
         String messageHeader,
 
-        int messageSender,
+        long messageSender,
 
-        int receiver,
+        long receiver,
 
         String receiverGroup,
